deduplicate logic shared between DataExportContainer and UsageSnapshotsContainer into shared parent container
