+++ services/QuillLMS/app/assets/stylesheets/pages/premium_hub/data-export.scss
@@ -1,124 +1,137 @@
-.data-export-page-container {
-  .filter-container {
-    .filters {
-      .dropdown {
-        margin-right: 0px !important;
-      }
-    }
+.data-export-main {
+  padding: 32px 0px 32px 64px;
+  max-width: calc(100% - 317px);
+
+  .header {
+    padding-right: 64px;
+    max-width: 1278px;
   }
-  .data-export-main {
-    padding: 32px 0px 32px 64px;
-    max-width: calc(100% - 317px);
-    .header {
-      padding-right: 64px;
-      max-width: 1278px;
+
+  .data-export-container {
+    display: flex;
+    flex-direction: column;
+
+    section {
+      padding: 0;
     }
-    .data-export-container {
-      display: flex;
-      flex-direction: column;
-      section {
-        padding: 0;
+
+    .fields-section,
+    .preview-section {
+      h3 {
+        font-size: 20px;
+        font-weight: 600;
+        color: $quill-grey-90;
       }
-      .fields-section, .preview-section {
-        h3 {
-          font-size: 20px;
-          font-weight: 600;
-          color: $quill-grey-90;
-        }
-      }
-      .fields-section {
-        margin-top: 32px;
-        max-width: 1278px;
-        padding-right: 64px;
-        .fields-container {
+    }
+
+    .fields-section {
+      margin-top: 32px;
+      max-width: 1278px;
+      padding-right: 64px;
+
+      .fields-container {
+        display: flex;
+        flex-wrap: wrap;
+        margin: 16px 0px 32px 0px;
+
+        .checkbox-container {
           display: flex;
-          flex-wrap: wrap;
-          margin: 16px 0px 32px 0px;
-          .checkbox-container {
-            display: flex;
-            flex: 1 0 16%;
-            max-height: 35px;
-            max-width: 165px;
-            min-width: 165px;
-            padding: 8px;
-            align-items: center;
-            gap: 8px;
-            border-radius: 8px;
-            border: 1px solid $quill-grey-5;
-            background: $quill-grey-1;
-            margin: 0px 8px 8px 0px;
-            label {
-              font-weight: 400;
-              margin-bottom: 0;
-            }
-          }
-        }
-      }
-      .preview-section {
-        padding-right: 64px;
-        .preview-disclaimer-container {
-          display: flex;
-          padding: 16px;
+          flex: 1 0 16%;
+          max-height: 35px;
+          max-width: 165px;
+          min-width: 165px;
+          padding: 8px;
+          align-items: center;
+          gap: 8px;
           border-radius: 8px;
           border: 1px solid $quill-grey-5;
-          background: #F4F4F4;
-          margin: 16px 0px 24px 0px;
-          img {
-            margin-right: 10px;
+          background: $quill-grey-1;
+          margin: 0px 8px 8px 0px;
+
+          label {
+            font-weight: 400;
+            margin-bottom: 0;
           }
         }
       }
-      .data-export-table {
-        border: 1px solid $quill-grey-5;
+    }
+
+    .preview-section {
+      padding-right: 64px;
+
+      .preview-disclaimer-container {
+        display: flex;
+        padding: 16px;
         border-radius: 8px;
-        overflow-y: clip;
-        overflow-x: auto;
-        width: calc(100% - 64px);
-        .data-table-headers {
-          background-color: $quill-grey-1;
+        border: 1px solid $quill-grey-5;
+        background: #F4F4F4;
+        margin: 16px 0px 24px 0px;
+
+        img {
+          margin-right: 10px;
+        }
+      }
+    }
+
+    .data-export-table {
+      border: 1px solid $quill-grey-5;
+      border-radius: 8px;
+      overflow-y: clip;
+      overflow-x: auto;
+      width: calc(100% - 64px);
+
+      .data-table-headers {
+        background-color: $quill-grey-1;
+        min-height: 52px;
+
+        th {
+          color: $quill-grey-90;
+          font-weight: 700;
+          align-items: center;
           min-height: 52px;
-          th {
-            color: $quill-grey-90;
-            font-weight: 700;
+          max-width: 152px;
+
+          &:not(:last-child) {
+            border-right: 1px solid $quill-grey-5;
+          }
+        }
+      }
+
+      .data-table-body {
+        .data-table-row {
+          td {
+            display: flex;
             align-items: center;
             min-height: 52px;
             max-width: 152px;
+            white-space: break-spaces;
+            padding-right: 24px;
+            margin-right: 24px !important;
+            font-size: 12px;
+
             &:not(:last-child) {
               border-right: 1px solid $quill-grey-5;
             }
-          }
-        }
-        .data-table-body {
-          .data-table-row {
-            td {
-              display: flex;
-              align-items: center;
-              min-height: 52px;
-              max-width: 152px;
-              white-space: break-spaces;
-              padding-right: 24px;
-              margin-right: 24px !important;
-              font-size: 12px;
-              &:not(:last-child) {
-                border-right: 1px solid $quill-grey-5;
-              }
-              .quill-tooltip-trigger {
-                span {
-                  padding-right: 24px;
-                }
+
+            .quill-tooltip-trigger {
+              span {
+                padding-right: 24px;
               }
             }
-            td:first-of-type {
+          }
+
+          td:first-of-type {
+            .quill-tooltip {
+              right: 0;
+            }
+          }
+
+          td:last-of-type {
+            .quill-tooltip-wrapper {
+              left: auto !important;
+
               .quill-tooltip {
                 right: 0;
-              }
-            }
-            td:last-of-type {
-              .quill-tooltip-wrapper {
-                left: auto !important;
-                .quill-tooltip {
-                  right: 0;
-                }
               }
             }
           }
