+++ services/QuillLMS/app/assets/stylesheets/pages/premium_hub/usage-snapshot-report.scss
@@ -1,4 +1,4 @@
-#premium-hub .usage-snapshots-container, .data-export-page-container {
+#premium-hub .filterable-reports-container {
   display: flex;
   min-height: 100vh;
   .dropdown {
