+++ services/QuillLMS/client/app/bundles/PremiumHub/components/usage_snapshots/shared.ts
@@ -316,10 +316,3 @@ export const snapshotSections = [
     ]
   }
 ]
-
-export function selectionsEqual(firstSelection, secondSelection) {
-  return (
-    firstSelection == secondSelection || // less strict comparison so that undefined and null are treated as equal
-    unorderedArraysAreEqual(firstSelection, secondSelection)
-  )
-}
